changed from float to BigDecimal for money value everywhere

+++ src/be/Ticket.java
@@ -1,19 +1,21 @@
 package be;
 
+import java.math.BigDecimal;
+
 public class Ticket {
 
     private int id;
     private String eventName;
     private String ticketType;
     private int quantity;
-    private float ticketPrice;
+    private BigDecimal ticketPrice;
     private String startDate;
     private String customerName;
     private String customerEmail;
     private String UUID;
 
 
-    public Ticket(int id, String eventName, String ticketType, float ticketPrice,  String startDate, String customerName, String customerEmail, String UUID, int quantity) {
+    public Ticket(int id, String eventName, String ticketType, BigDecimal ticketPrice,  String startDate, String customerName, String customerEmail, String UUID, int quantity) {
         this.id = id;
         this.eventName = eventName;
         this.ticketType = ticketType;
@@ -25,21 +27,21 @@ public class Ticket {
         this.quantity = quantity;
     }
 
-    public Ticket (String nameEvent, String ticketType,  String startDate, float price){
+    public Ticket (String nameEvent, String ticketType,  String startDate, BigDecimal price){
         this.eventName = nameEvent;
         this.ticketType = ticketType;
         this.startDate = startDate;
         this.ticketPrice = price;
     }
 
-    public Ticket (int id, String ticketType, int quantity, float ticketPrice) {
+    public Ticket (int id, String ticketType, int quantity, BigDecimal ticketPrice) {
         this.id = id;
         this.ticketType = ticketType;
         this.quantity = quantity;
         this.ticketPrice = ticketPrice;
     }
 
-    public Ticket (String ticketType, int quantity, float ticketPrice) {
+    public Ticket (String ticketType, int quantity, BigDecimal ticketPrice) {
         this.ticketType = ticketType;
         this.quantity = quantity;
         this.ticketPrice = ticketPrice;
@@ -77,11 +79,11 @@ public class Ticket {
         this.ticketType = ticketType;
     }
 
-    public float getTicketPrice() {
+    public BigDecimal getTicketPrice() {
         return ticketPrice;
     }
 
-    public void setTicketPrice(float ticketPrice) {
+    public void setTicketPrice(BigDecimal ticketPrice) {
         this.ticketPrice = ticketPrice;
     }
